Use set-param! for paramter W:
This is a workaround to circumvent a "bug" which was introduced with
the advent of the materials library `materials.scm` in Meep v1.6.0.
This library predefines some variables (material constants). We have
naming clash with tungsten (W) and the width of the integration window (W).
Once a vaiable is already defined, define-param can no longer be used.